backend/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared libraries for Project Management Platform microservices"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
